requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==4.2.5
django-anymail==10.1
django-celery-beat==2.5.0
django-redis==5.3.0
django-timezone-field==6.0.1
djangorestframework==3.14.0
factory-boy==3.3.0
Faker==19.6.1
flake8==6.1.0
idna==3.4
iniconfig==2.0.0
kombu==5.3.2
mailjet-rest==1.3.4
mccabe==0.7.0
packaging==23.1
pluggy==1.3.0
prompt-toolkit==3.0.39
psycopg2==2.9.7
pycodestyle==2.11.0
pyflakes==3.1.0
pytest==7.4.2
pytest-django==4.5.2
pytest-mock==3.11.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-http-client==3.3.7
pytz==2023.3.post1
redis==5.0.0
requests==2.31.0
sendgrid==6.10.0
six==1.16.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
